+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Ics.py changelog
  - Adds support for Attendees to Events
  - Added LAST-MODIFIED attribute support
  - Event equality now checks all fields (except uid)
+ - alarms in Event and Todo are now consistently lists and not a mix between set() and list()
 
 **************
 0.4